+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Added
+
+- Support Go 1.19.
+  Include compatibility testing and document support. (#3077)
+
 ## [1.9.0/0.0.3] - 2022-08-01
 
 ### Added